Rename `Force` to `ForceSimulation`
