+++ packages/layerchart/src/lib/components/ForceSimulation.svelte
